Restructure of entire project structure. API stuff is now in 'pkg/api', and there is no longer a 'utils/helper.go' (although this can be added back if need be). All code that relied on previous structure was changed to accomodate. 'main.go' was also restructured to work as a main server application.
