+++ Dockerfile
@@ -7,15 +7,14 @@ COPY go.mod .
 COPY go.sum .
 RUN go mod download
 
-
 # https://docs.docker.com/build/cache/optimize/
 COPY . .
 
-# RUN --mount=type=cache,target=GOCACHE \
-#     go build -o the_cookie_jar
-
 RUN --mount=type=cache,target=GOCACHE \
-    go build cmd/client/server.go
+     go build -o the_cookie_jar
+
+# RUN --mount=type=cache,target=GOCACHE \
+#     go build cmd/client/server.go
 
 FROM alpine
 WORKDIR /root
@@ -27,4 +26,4 @@ ADD https://github.com/ufoscout/docker-compose-wait/releases/download/2.2.1/wait
 RUN chmod +x /wait
 
 
-CMD /wait && ./server
+CMD /wait && ./the_cookie_jar
