+++ README.md
@@ -1,26 +1,39 @@
-### CSE 312 Group Project 
+CSE 312 Group Project 
 
-# the_cookie_jar API
+## the_cookie_jar
 
-### the_cookie_jar is a Learning Management System designed to be both modular and extensible, using a plugin based architecture
+### Description
 
+A Learning Management System (LMS) designed to have ease-of-use and basic professor/student interactions.
 
-clone this repo
+### Running the Server
 
-create file [.env] with the following variables set for local developement:
+1. Clone this repo, go to branch of choice (`main` is production, `dev` is close to production).
+2. Create an environment file [`.env`] in root directory with contents:
 ```
-  MONGODB_URI=mongodb://database:27017/
-  DB_NAME=the_cookie_jar
+MONGODB_URI=mongodb://database:27017/
+DB_NAME=the_cookie_jar
 ```
-
-from project directory: 
-
-  ```
-    clear; docker compose up --build --force-recreate
-  ```
-
-EXPECT:
+3. Run server with Docker from base directory with command `docker compose up --build --force-recreate` and expect:
 ```
-cookie_lms  | >> [db] Pinged your deployment. You successfully connected to MongoDB! mongodb://database:27017/ 
-cookie_lms  | [GIN] 2024/10/02 - 22:04:21 | 400 |   20.532541ms |             ::1 | POST     "/users/maker_name"
+cookie_lms  | Checking availability of database:27017
+cookie_lms  | ...
+cookie_lms  | Host database:27017 is now available
+cookie_lms  | ...
+cookie_lms  | -----------------------------------
+cookie_lms  | the_cookie_jar server is running...
+cookie_lms  | -----------------------------------
+cookie_lms  | ...
 ```
+4. You may now go to `http://localhost:8080/` on your webbrowser and interact with the server.
+
+### Folder Layout
+
+| Folder | Description |
+|--------|-------------|
+| cmd | Applications that make use of `pkg` libraries.
+| docs | Any documentation related to the project.
+| pkg | Our libraries used in the project.
+| public | Files to send to clients based on requests (i.e. the webpage, CSS, etc.)
+
+*Note: `main.go` is the only application not located in `cmd` folder, this is because it is the **main** server application that is expected to interact with clients for production purposes.*
