+++ docker-compose.yml
@@ -3,7 +3,7 @@ services:
     container_name: "cookie_lms"
     build: .
     ports:
-      - '8080:6669'
+      - '8080:8080'
     environment:
       WAIT_HOSTS: database:27017
       GOCACHE: /.go-build/
