+++ main.go
@@ -2,39 +2,26 @@ package main
 
 import (
 	"fmt"
-	"os"
-	"sync"
-	"time"
 
-	"github.com/danmuck/the_cookie_jar/api/models"
+	"github.com/danmuck/the_cookie_jar/pkg/api"
+	"github.com/gin-gonic/gin"
 )
 
 func main() {
+	// Want Gin debug statements? Comment this line out.
+	gin.SetMode(gin.ReleaseMode)
 
-	new_user := models.NewUser("Big", "password")
-	var tmp int = 12
+	router := api.BaseRouter()
+	fmt.Println("-----------------------------------")
+	fmt.Println("the_cookie_jar server is running...")
+	fmt.Println("-----------------------------------")
+	err := router.Run(":8080")
 
-	fmt.Fprintf(os.Stdout, "\n\nSup, %s --%s age: %d \n\n", new_user.GetUsername(), new_user.GetStatus_String(), tmp)
-	fmt.Println("Using goroutines, iterate the bytes of [new_user.id] and print them ..\n .. they are indexed in the order the goroutines were created .. ")
-
-	var wg sync.WaitGroup
-	wg.Add(len(new_user.ID))
-	for i, id := range new_user.GetId() {
-		go func(wg *sync.WaitGroup) {
-			// by calling a function using [go nameOfFunction()] a goroutine is created
-			// goroutines run concurrently
-			time.Sleep(2500 * time.Millisecond)
-			fmt.Printf("%d: %d \n", i, id)
-			wg.Done()
-		}(&wg)
+	// If a general error in the server occurs
+	if err != nil {
+		panic(fmt.Errorf("%v", err))
 	}
-	wg.Wait()
-
-	// sleep to wait for the goroutines to finish
-	// time.Sleep(2 * time.Second)
-	fmt.Println("new_user_id: ", new_user.GetId())
-	fmt.Println("\n\n\n\n\n!! >> See Dockerfile CMD \n\n ")
-	fmt.Println("\n\n!! >>  go build /cmd/client/server.go \n\n ")
-	fmt.Println("\n\n!! >> PLACEHOLDER (main.go) !! \n\n ")
-
+	fmt.Println("-----------------------------------")
+	fmt.Println("the_cookie_jar server is closing...")
+	fmt.Println("-----------------------------------")
 }
