--- plugins/dev/models.go
@@ -1,1 +1,0 @@
-package dev
