--- plugins/tmp/controllers.go
@@ -1,1 +1,0 @@
-package tmp
