--- plugins/tmp/templates/index.tmpl
@@ -1,14 +1,0 @@
-<html>
-	<h1>Template Demo</h1>
-
-	<h2>
-		{{ .title }}
-	</h2>
-	<hr>
-	<h3>
-		{{ .sub_title }}
-	</h3>
-	<b style="color: red">
-		{{ .body }}
-	</b>
-</html>
